Match func_8003C0A4
Allocator-init wrapper, sibling of func_8003C420. Allocates a Struct3CAF0
via func_80071408, stores the pointer in D_8010ADDC, then sets unk1C/unk1E
(s16), unk26 (u8), and the high byte of the s16 at 0x24 (the byte at 0x25,
written via ((u8*)&unk24)[1] matching the idiom used in func_8003BEF0).
Also retyped the call-site externs in 33680.c/35E20.c/37FE0.c to the
correct (s16,s16,u8,u8) signature.

+++ src/33680.c
@@ -7,7 +7,7 @@ extern void func_80034BC4();
 extern s32 D_8010B1C0;
 
 extern void func_800373AC(s32 *arg0);
-extern void func_8003C0A4(s32 arg0, s32 arg1, s32 arg2, s32 arg3);
+extern void func_8003C0A4(s16 arg0, s16 arg1, u8 arg2, u8 arg3);
 extern void func_80041DD4(s32 arg0, s32 arg1);
 extern s32 func_80041FB4(s32 arg0);
 extern void func_80042034(s32 arg0);
